Logger moved into /lib folder
