+++ src/service/cli/version.js
@@ -2,7 +2,7 @@
 
 const packageJsonFile = require(`../../../package.json`);
 const chalk = require(`chalk`);
-const { getLogger } = require('../logger.js')
+const { getLogger } = require('../lib/logger.js')
 const logger = getLogger()
 
 module.exports = {
